+++ src/components/QueryForm.vue
@@ -45,8 +45,18 @@
             </sui-form-field>
 
             <!-- fetch -->
-            <sui-form-field v-if="queryType === 'fetch'">
-                <label>URL</label>
+            <sui-form-field v-if="queryType === 'fetch' || queryType === 'sparql'">
+                <label style="float:left">URL</label>
+                <sui-dropdown text="Common SPARQL Services" pointing right style="float:right" v-if="queryType === 'sparql'">
+                    <sui-dropdown-menu>
+                        <sui-dropdown-item v-for="item in sparqlExamples" :key="item.key" @click.prevent="setUrl(item)">
+                            {{ item.text }}
+                            <!-- <span class="description" v-html="item.value" /> -->
+                        </sui-dropdown-item>
+                    </sui-dropdown-menu>
+                </sui-dropdown>
+
+
                 <sui-input
                     type="url"
                     placeholder="URL"
@@ -55,6 +65,45 @@
                     v-model="url"
                 />
             </sui-form-field>
+
+            <!-- sparql params? -->
+            <sui-segment v-if="queryType === 'fetch' || queryType === 'sparql'">
+                <p><strong>Request Headers</strong></p>
+
+                <sui-grid :columns="2">
+                    <sui-grid-row>
+                        <sui-grid-column>
+                            <sui-form-field>
+                                <sui-checkbox v-model="setAcceptHeader" label="Set Accept Header" />
+                            </sui-form-field>
+                        </sui-grid-column>
+                        <sui-grid-column>
+                            <sui-form-field>
+                                <label>Authorization</label>
+
+                                <sui-dropdown
+                                    v-model="authorizationType"
+                                    fluid
+                                    :options="authorizationTypes"
+                                    selection
+                                />
+
+                                <sui-input
+                                    v-if="authorizationType != 'none'"
+                                    v-model="username"
+                                    :placeholder="authorizationType == 'Basic' ? 'Username' : 'Token'"
+                                />
+
+                                <sui-input
+                                    v-if="authorizationType == 'Basic'"
+                                    v-model="password"
+                                    placeholder="Password"
+                                />
+                            </sui-form-field>
+                        </sui-grid-column>
+                    </sui-grid-row>
+                </sui-grid>
+            </sui-segment>
 
              <sui-form-field v-if="showDisplayAs">
                 <label>Display Results As</label>
@@ -85,7 +134,7 @@
 </template>
 
 <script>
-import { queryTypes, formats, outputFormats } from '../constants';
+import { queryTypes, formats, outputFormats, sparqlExamples } from '../constants';
 import CypherComponent from '@/components/CypherComponent'
 
 export default {
@@ -116,10 +165,18 @@ export default {
         queryTypes,
         formats,
         displayAsFixed: '',
+
+        // SPARQL options
+        sparqlExamples,
+        authorizationTypes: [
+            { key: 'none', text: 'None', value: 'none',  },
+            { key: 'basic', text: 'Basic', value: 'Basic',  },
+            { key: 'bearer', text: 'Bearer', value: 'Bearer',  },
+        ],
     }),
     computed: {
         showTextbox() {
-            return (this.queryType === 'inline' || this.queryType === 'local' && this.value)
+            return (this.queryType === 'inline' || this.queryType === 'local' && this.value) || this.queryType === 'sparql'
         },
         cypher() {
             const displayAs = this.showDisplayAs ? `.${this.displayAs}` : ''
@@ -127,19 +184,43 @@ export default {
             const procedure = `${this.procedure}${displayAs}.${this.procedureType}`
             let value = this.queryType === 'fetch' ? this.url : `\n${this.value.replace(/'/g, "\\'")}\n`
 
+            if ( this.queryType === 'sparql' ) {
+                value = `${this.url}${encodeURI(this.value)}`
+            }
+
             return `CALL ${procedure}(
   '${value}',
-  '${this.format}'
+  '${this.format}' ${this.parameters}
 )`
         },
         procedureType() {
-            return  this.queryType === 'fetch' ? 'fetch' : 'inline'
+            return  this.queryType === 'fetch' || this.queryType === 'sparql' ? 'fetch' : 'inline'
         },
         runDisabled() {
             return ( this.queryType === 'fetch' && !this.url ) || ( this.queryType !== 'fetch' && this.value === '' )
         },
         accept() {
             return [ 'text/*', ].concat(  formats.find(f => f.value === this.format).types )
+        },
+        acceptHeader() {
+            return formats.find(f => f.value === this.format).types[0]
+        },
+        authorizationHeader() {
+            return this.authorizationType === 'Basic'
+                ? `Authorization: "Basic ${btoa(`${this.username}:${this.password}`)}"`
+                : `Authorizaion: "Bearer ${this.username}"`
+        },
+        parameters() {
+            const headerParams = []
+            if ( this.setAcceptHeader ) {
+                headerParams.push(`Accept: "${this.acceptHeader}"`)
+            }
+
+            if ( this.authorizationType !== 'none' ) {
+                headerParams.push(this.authorizationHeader)
+            }
+
+            return headerParams.length ? `,\n  { headerParams: { ${headerParams.join(', ')} } }` : '';
         },
         queryType: {
             get() {
@@ -184,6 +265,46 @@ export default {
                 return this.$store.commit(mutation, value)
             },
         },
+        setAcceptHeader: {
+            get() {
+                return this.$store.state.setAcceptHeader
+            },
+            set(value) {
+                return this.$store.commit('setAcceptHeader', value)
+            },
+        },
+        authorization: {
+            get() {
+                return this.$store.state.authorization
+            },
+            set(value) {
+                return this.$store.commit('setAuthorization', value)
+            },
+        },
+        authorizationType: {
+            get() {
+                return this.$store.state.authorizationType
+            },
+            set(value) {
+                return this.$store.commit('setAuthorizationType', value)
+            },
+        },
+        username: {
+            get() {
+                return this.$store.state.username
+            },
+            set(value) {
+                return this.$store.commit('setUsername', value)
+            },
+        },
+        password: {
+            get() {
+                return this.$store.state.password
+            },
+            set(value) {
+                return this.$store.commit('setPassword', value)
+            },
+        },
     },
     methods: {
         fileChanged(e) {
@@ -193,7 +314,14 @@ export default {
             }
 
             fr.readAsText(e.target.files[0])
-        }
+        },
+        setUrl(selected) {
+            this.url = selected.url
+
+            if ( this.queryType === 'sparql' && selected.query && this.value === '' ) this.value = selected.query
+            if ( selected.setAcceptHeader !== undefined ) this.setAcceptHeader = selected.setAcceptHeader
+            if ( selected.authorizationType !== undefined ) this.authorizationType = selected.authorizationType
+        },
     },
     watch: {
         queryType() {