+++ release-notes.md
@@ -1,3 +1,9 @@
+# n10s version 0.2.10
+
+- Added a new `SPARQL Query` option to the RDF source dropdown to allow you to copy and paste a SPARQL query and have n10s pull the results for you
+- Added list of Common SPARQL services to prepopulate the URL field
+- Added a **Request Headers** section to the import form to allow you to set the `Accept:` and `Authorization:` headers
+
 # n10s version 0.2.9
 
 - Fixed bug on export screen where URL would always be `localhost`